+++ pgcrypto_fields/__init__.py
@@ -7,4 +7,4 @@ HMAC_SQL = "hmac(%s, '{}', 'sha512')".format(settings.PGCRYPTO_KEY)
 PGP_PUB_ENCRYPT_SQL = "pgp_pub_encrypt(%s::text, dearmor('{}'))".format(
     settings.PUBLIC_PGP_KEY,
 )
-PGP_SYM_ENCRYPT_SQL = "pgp_sym_encrypt(%s, '{}')".format(settings.PGCRYPTO_KEY)
+PGP_SYM_ENCRYPT_SQL = "pgp_sym_encrypt(%s::text, '{}')".format(settings.PGCRYPTO_KEY)