+++ tests/factories.py
@@ -13,4 +13,5 @@ class EncryptedTextFieldModelFactory(factory.DjangoModelFactory):
 
     integer_pgp_pub_field = 42
     pgp_pub_field = factory.Sequence('Text with public key {}'.format)
+    integer_pgp_sym_field = 43
     pgp_sym_field = factory.Sequence('Text with symmetric key {}'.format)