+++ tests/test_fields.py
@@ -39,6 +39,7 @@ class TestEncryptedTextFieldModel(TestCase):
             'hmac_field',
             'integer_pgp_pub_field',
             'pgp_pub_field',
+            'integer_pgp_sym_field',
             'pgp_sym_field',
         )
         self.assertCountEqual(fields, expected)
@@ -52,6 +53,7 @@ class TestEncryptedTextFieldModel(TestCase):
         self.assertIsInstance(instance.hmac_field, str)
         self.assertIsInstance(instance.integer_pgp_pub_field, int)
         self.assertIsInstance(instance.pgp_pub_field, str)
+        self.assertIsInstance(instance.integer_pgp_sym_field, int)
         self.assertIsInstance(instance.pgp_sym_field, str)
 
     def test_fields_descriptor_is_not_instance(self):